requirements.txt
Django>=4.2
djangorestframework
requests
python-dotenv
django-cors-headers
